+++ starkinfra/individualdocument/log/__init__.py
@@ -1,1 +1,0 @@
-from .__log import query, page, get